Moved contents of src/lmnt to src
