+++ LMNT/src/archive.c
@@ -7,7 +7,7 @@
 #include <limits.h>
 
 #include "lmnt/extcalls.h"
-#include "lmnt/helpers.h"
+#include "helpers.h"
 
 
 //
