+++ LMNT/src/extcalls.c
