+++ LMNT/src/helpers.h
