+++ LMNT/src/jit.c
@@ -2,7 +2,7 @@
 #include "lmnt/interpreter.h"
 #include "lmnt/jit.h"
 #include "lmnt/platform.h"
-#include "lmnt/jit/hosthelpers.h"
+#include "jit/hosthelpers.h"
 
 #if defined(LMNT_JIT_HAS_X86_64)
 lmnt_result lmnt_jit_x86_64_compile(lmnt_ictx* ctx, const lmnt_def* def, lmnt_jit_fn_data* fndata, lmnt_jit_compile_stats* stats);
