+++ LMNT/src/opcodes.c
