+++ LMNT/src/ops_math.c
