+++ LMNT/src/ops_trig.c
