+++ LMNT/src/ops_util.c
