+++ LMNT/src/validation.c
@@ -1,5 +1,5 @@
 #include "lmnt/validation.h"
-#include "lmnt/helpers.h"
+#include "helpers.h"
 
 #include <string.h>
